requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.0
python-dotenv>=1.0.0
